requirements.txt
numpy==1.20.1
opencv_python==4.5.1.48
keyboard==0.13.5
PyAutoGUI==0.9.52
Pillow==8.1.2